+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.20.1] - 2025-05-08
+- Added node control functions:
+  - control(action) - Sends 'reset' or 'rebuild' commands to node
+  - rebuild() - Shortcut for control('rebuild')
+  - reset() - Shortcut for control('reset')
+
 ## [2.20.0] - 2025-05-06
 -   Added OrganizationMembers Class and Endpoint SDK